+++ anubis-core/src/main/java/com/anubis/core/dao/DHT11Repo.java
@@ -1,0 +1,8 @@
+package com.anubis.core.dao;
+
+
+import com.anubis.core.entity.sensor.DHT11Data;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DHT11Repo extends JpaRepository<DHT11Data, Long> {
+}
