+++ anubis-core/src/main/java/com/anubis/core/entity/interfaces/Created.java
@@ -1,0 +1,18 @@
+package com.anubis.core.entity.interfaces;
+
+import com.anubis.core.entity.family.FamilyMember;
+
+import java.time.LocalDateTime;
+
+/**
+ * The Created interface is for any entity that is to have a
+ * CreateBy (FamilyMember) and CreatedAt (LocalDateTime)
+ */
+public interface Created {
+
+    FamilyMember getCreatedBy();
+    void setCreatedBy(FamilyMember familyMember);
+
+    LocalDateTime getCreatedAt();
+    void setCreatedAt(LocalDateTime createdAt);
+}
