+++ anubis-core/src/main/java/com/anubis/core/entity/interfaces/Discriminated.java
@@ -1,0 +1,7 @@
+package com.anubis.core.entity.interfaces;
+
+public interface Discriminated {
+
+    String getDiscriminator();
+
+}
