+++ anubis-core/src/main/java/com/anubis/core/entity/interfaces/Named.java
@@ -1,0 +1,8 @@
+package com.anubis.core.entity.interfaces;
+
+public interface Named {
+
+    String getName();
+    void setName(String name);
+
+}
