+++ anubis-core/src/main/java/com/anubis/core/entity/interfaces/Reminded.java
@@ -1,0 +1,14 @@
+package com.anubis.core.entity.interfaces;
+
+import com.anubis.core.entity.family.Reminder;
+
+import java.util.List;
+
+/**
+ * The Task interface is the base for all tasks.
+ */
+public interface Reminded {
+    List<Reminder> getReminders();
+    void setReminders(List<Reminder> reminders);
+    void addReminder(Reminder reminder);
+}
