+++ anubis-core/src/main/java/com/anubis/core/entity/sub/Imported.java
@@ -1,0 +1,10 @@
+package com.anubis.core.entity.sub;
+
+import java.util.Date;
+
+public interface Imported {
+
+    Date getImportDate();
+
+    void setImportDate(Date importDate);
+}
