+++ anubis-core/src/main/java/com/anubis/core/exception/AnubisPythonExecutionException.java
@@ -1,0 +1,13 @@
+package com.anubis.core.exception;
+
+import org.apache.commons.exec.ExecuteException;
+
+public class AnubisPythonExecutionException extends ExecuteException {
+    public AnubisPythonExecutionException(String message, int exitValue) {
+        super(message, exitValue);
+    }
+
+    public AnubisPythonExecutionException(String message, int exitValue, Throwable cause) {
+        super(message, exitValue, cause);
+    }
+}
