--- anubis-core/src/main/java/com/tcramer/anubis/core/dao/DHT11Repo.java
@@ -1,8 +1,0 @@
-package com.tcramer.anubis.core.dao;
-
-
-import com.tcramer.anubis.core.entity.sensor.DHT11Data;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface DHT11Repo extends JpaRepository<DHT11Data, Long> {
-}
