--- anubis-core/src/main/java/com/tcramer/anubis/core/entity/family/DailyTask.java
@@ -1,23 +1,0 @@
-package com.tcramer.anubis.core.entity.family;
-
-import com.tcramer.anubis.core.constants.TaskType;
-
-import javax.persistence.DiscriminatorValue;
-import javax.persistence.Entity;
-
-@Entity
-@DiscriminatorValue(value= TaskType.Discriminator.DAILY)
-public class DailyTask extends Task {
-
-    public DailyTask(String name) {
-        super(name);
-    }
-
-    public DailyTask() {
-    }
-
-    @Override
-    public String getDiscriminator() {
-        return TaskType.Discriminator.DAILY;
-    }
-}
