--- anubis-core/src/main/java/com/tcramer/anubis/core/entity/family/WeeklyTask.java
@@ -1,16 +1,0 @@
-package com.tcramer.anubis.core.entity.family;
-
-import com.tcramer.anubis.core.constants.TaskType;
-
-import javax.persistence.DiscriminatorValue;
-import javax.persistence.Entity;
-
-@Entity
-@DiscriminatorValue(value= TaskType.Discriminator.WEEKLY)
-public class WeeklyTask extends Task {
-
-    @Override
-    public String getDiscriminator() {
-        return TaskType.Discriminator.WEEKLY;
-    }
-}
