--- anubis-core/src/main/java/com/tcramer/anubis/core/entity/interfaces/Discriminated.java
@@ -1,7 +1,0 @@
-package com.tcramer.anubis.core.entity.interfaces;
-
-public interface Discriminated {
-
-    String getDiscriminator();
-
-}
