--- anubis-core/src/main/java/com/tcramer/anubis/core/entity/interfaces/Named.java
@@ -1,8 +1,0 @@
-package com.tcramer.anubis.core.entity.interfaces;
-
-public interface Named {
-
-    String getName();
-    void setName(String name);
-
-}
