--- anubis-core/src/main/java/com/tcramer/anubis/core/entity/sub/Imported.java
@@ -1,10 +1,0 @@
-package com.tcramer.anubis.core.entity.sub;
-
-import java.util.Date;
-
-public interface Imported {
-
-    Date getImportDate();
-
-    void setImportDate(Date importDate);
-}
