+++ anubis-home-api/src/main/java/com/tcramer/anubis/home/service/DHT11Service.java
@@ -1,6 +1,6 @@
 package com.tcramer.anubis.home.service;
 
-import com.tcramer.anubis.core.entity.sensor.DHT11Data;
+import com.anubis.core.entity.sensor.DHT11Data;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
