+++ anubis-home-api/src/main/java/com/tcramer/anubis/home/service/DHT11ServiceImpl.java
@@ -1,8 +1,8 @@
 package com.tcramer.anubis.home.service;
 
-import com.tcramer.anubis.core.dao.DHT11Repo;
-import com.tcramer.anubis.core.entity.sensor.DHT11Data;
-import com.tcramer.anubis.core.util.PythonExecUtil;
+import com.anubis.core.dao.DHT11Repo;
+import com.anubis.core.entity.sensor.DHT11Data;
+import com.anubis.core.util.PythonExecUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
